requirements.txt
pyTelegramBotAPI==4.12.0
yt-dlp==2025.8.18
python-dotenv==1.1.1
ffmpeg-python==0.2.1